requirements.txt
# LoanWise Pro - Dependencies
# ============================
# Install all packages with: pip install -r requirements.txt

# ==============================================
# Core Data Science
# ==============================================
numpy>=1.21.0
pandas>=1.3.0

# ==============================================
# Machine Learning
# ==============================================
scikit-learn>=1.0.0

# ==============================================
# Explainability (XAI)
# ==============================================
shap>=0.41.0

# ==============================================
# Web Dashboard (Streamlit)
# ==============================================
streamlit>=1.28.0

# ==============================================
# Visualization
# ==============================================
plotly>=5.10.0

# ==============================================
# Database - PostgreSQL
# ==============================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# ==============================================
# FastAPI Backend
# ==============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
python-multipart>=0.0.6

# ==============================================
# Authentication & Security
# ==============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# ==============================================
# Configuration & Validation
# ==============================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ==============================================
# HTTP Client (for external APIs)
# ==============================================
httpx>=0.26.0

# ==============================================
# Email (for notifications)
# ==============================================
aiosmtplib>=3.0.0
email-validator>=2.1.0

# ==============================================
# Rate Limiting & Background Tasks
# ==============================================
celery>=5.3.0

# ==============================================
# Logging & Monitoring
# ==============================================
structlog>=24.1.0
psutil>=5.9.0

# ==============================================
# Testing
# ==============================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# ==============================================
# Development Tools
# ==============================================
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
